+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New table output layout SolidTable.
 - Allow to use table layout and separator options for other commands.
+- Add raw formatter "raw_fmt".
 
 ### Fixed